Backend/auction_project/requirements.txt
asgiref==3.5.2
bleach==5.0.1
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-api-logger==1.1.11
Pillow==9.3.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
webencodings==0.5.1
django-phonenumber-field==7.0.0
mysqlclient==2.1.1

